+++ compass/api/__init__.py
@@ -1,0 +1,1 @@
+# This directory is a Python package.
