+++ compass/api/plugins/redis.py
