+++ compass/api/routes/authentication.py
@@ -5,13 +5,13 @@ from fastapi import APIRouter
 from fastapi import Depends
 from fastapi.security import OAuth2PasswordRequestForm
 
-from src.api.plugins.redis import depends_redis
-from src.api.schemas.auth import Token
-from src.api.utility.oauth2 import ACCESS_TOKEN_EXPIRE_MINUTES
-from src.api.utility.oauth2 import authenticate_user
-from src.api.utility.oauth2 import create_access_token
-from src.api.utility.oauth2 import custom_bearer_auth_exception
-from src.api.utility.oauth2 import store_session
+from api.plugins.redis import depends_redis
+from api.schemas.auth import Token
+from api.utility.oauth2 import ACCESS_TOKEN_EXPIRE_MINUTES
+from api.utility.oauth2 import authenticate_user
+from api.utility.oauth2 import create_access_token
+from api.utility.oauth2 import custom_bearer_auth_exception
+from api.utility.oauth2 import store_session
 
 router = APIRouter()
