+++ compass/api/startup.sh
